requirements.txt
dateparser==0.7.1
dotmap==1.3.8
Faker==1.0.7
gitdb2==2.0.5
GitPython==2.1.11
numpy==1.16.4
pandas==0.24.2
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
regex==2019.6.8
six==1.12.0
smmap2==2.0.5
text-unidecode==1.2
tzlocal==1.5.1
